add more rpc support (#8)
* feat(rpc): add Accounts config support to SimulateTransaction

* feat(rpc): add new rpc getSignaturesForAddress

* feat(rpc): add new rpc getBlocksWithLimit

* feat(rpc): add new rpc getTransaction

* feat(rpc): add new rpc getBlocks

* feat(rpc): add new rpc getBlock

* feat(rpc): update get epoch info response type

* feat(rpc): add new rpc getBlockHeight

* feat(rpc): add new rpc getFirstAvailableBlock

* feat(rpc): add new rpc getGenesisHash

* feat(rpc): add new rpc get identity

* feat(rpc): add brief desc to rpc functions

+++ client/get_epoch_info.go
@@ -3,13 +3,14 @@ package client
 import "context"
 
 type GetEpochInfoResponse struct {
-	AbsoluteSlot int `json:"absoluteSlot"`
-	BlockHeight  int `json:"blockHeight"`
-	Epoch        int `json:"epoch"`
-	SlotIndex    int `json:"slotIndex"`
-	SlotsInEpoch int `json:"slotsInEpoch"`
+	AbsoluteSlot uint64 `json:"absoluteSlot"`
+	BlockHeight  uint64 `json:"blockHeight"`
+	Epoch        uint64 `json:"epoch"`
+	SlotIndex    uint64 `json:"slotIndex"`
+	SlotsInEpoch uint64 `json:"slotsInEpoch"`
 }
 
+// GetEpochInfo returns information about the current epoch
 func (s *Client) GetEpochInfo(ctx context.Context, commitment Commitment) (GetEpochInfoResponse, error) {
 	res := struct {
 		GeneralResponse
